requirements.txt
APScheduler~=3.11.0
Flask==3.1.0
Flask-APScheduler==1.13.1
python-dotenv==1.0.1
GitPython==3.1.44
gunicorn==23.0.0
requests-oauthlib==2.0.0
oauthlib~=3.2.2
requests~=2.32.3
adafruit-circuitpython-neopixel==6.3.15
